+++ hata/discord/activity/fields/tests/test__ActivityParty__magic.py
@@ -8,7 +8,7 @@ def test__ActivityParty__repr():
     Tests whether ``ActivityParty.__repr__`` works as intended.
     """
     field = ActivityParty(
-        id_ = 'plain',
+        party_id = 'plain',
         size = 6,
         max_ = 12,
     )
@@ -21,7 +21,7 @@ def test__ActivityParty__eq():
     Tests whether ``ActivityParty.__repr__`` works as intended.
     """
     fields = {
-        'id_': 'plain',
+        'party_id': 'plain',
         'size': 6,
         'max_': 12,
     }
@@ -31,7 +31,7 @@ def test__ActivityParty__eq():
     vampytest.assert_eq(field_original, field_original)
     
     for field_name, field_value in (
-        ('id_', 'trance'),
+        ('party_id', 'trance'),
         ('size', 0),
         ('max_', 0),
     ):
@@ -44,7 +44,7 @@ def test__ActivityParty__hash():
     Tests whether ``ActivityParty.__hash__`` works as intended.
     """
     field = ActivityParty(
-        id_ = 'plain',
+        party_id = 'plain',
         size = 6,
         max_ = 12,
     )
@@ -64,7 +64,7 @@ def test__ActivityParty__bool():
     
     
     for field_name, field_value in (
-        ('id_', 'trance'),
+        ('party_id', 'trance'),
         ('size', 6),
         ('max_', 12),
     ):